+++ modules/mysql/src/test/java/org/rnorth/testcontainers/junit/SimpleMySQLTest.java
