+++ modules/mysql/src/test/resources/somepath/init_mysql.sql
