+++ modules/postgresql/src/main/resources/META-INF/services/org.rnorth.testcontainers.containers.DatabaseContainer
@@ -1,0 +1,1 @@
+org.rnorth.testcontainers.containers.PostgreSQLContainer
